requirements.txt
flask
python-dotenv
openai
werkzeug
flask-cors